refactor: SelectUser and Forms
