+++ frontend/src/App.jsx
@@ -18,8 +18,6 @@ import CoursesTeacher from './pages/teacher/CoursesTeacher'
 import NotFound from './pages/NotFound'
 import RequireAuth from './components/RequireAuth'
 import DashboardLayout from './components/Layouts/DashboardLayout'
-import StudentConfigAccount from './components/studentConfigAccount'
-import TeacherConfigAccount from './components/TeacherConfigAccount'
 import CoursesList from './components/Teacher/CoursesList'
 import { alumnosNavConfig } from './pages/dashboard/layoutConfig'
 import { teacherNavConfig } from './pages/teacher/teacherConfig'
@@ -31,8 +29,6 @@ function App() {
       <Route path="/login" element={<Login />} />
       <Route path="/signup" element={<Signup />} />
       <Route path="/select-user-type" element={<SelectUserType />} />
-      <Route path="/config/teacher" element={<TeacherConfigAccount />} />
-      <Route path="/config/student" element={<StudentConfigAccount />} />
       <Route
         path="/dashboard"
         element={
